Fix user losing focus if they type really fast
We would move focus from the input in prefix based mode even though we were not yet in scan mode.

We also solve the problem in time based mode but with one catch.
There is still unsolvable case where the first char is entered in the input before scan mode is detected.
After changes in this commit, two chars will be entered within input until we detect scan mode.